feat: enhancements

+++ src/main/java/id/ac/ui/cs/advprog/mysawitbe/modules/pembayaran/application/dto/WalletBalanceDTO.java
@@ -9,6 +9,6 @@ import java.util.UUID;
  */
 public record WalletBalanceDTO(
         UUID userId,
-        int balance,
+        long balance,
         LocalDateTime lastUpdated
 ) {}
